Refine typing; mark TODOs
