+++ packages/back-end/types/stats.d.ts
@@ -1,0 +1,48 @@
+import type { MetricStats } from "./metric";
+
+interface BaseVariationResponse {
+  cr: number;
+  value: number;
+  users: number;
+  denominator?: number;
+  stats: MetricStats;
+  expected?: number;
+  uplift?: {
+    dist: string;
+    mean?: number;
+    stddev?: number;
+  };
+  ci?: [number, number];
+}
+
+interface BayesianVariationResponse extends BaseVariationResponse {
+  chanceToWin?: number;
+  risk?: [number, number];
+}
+
+interface FrequentistVariationResponse extends BaseVariationResponse {
+  pValue?: number;
+}
+
+interface BaseDimensionResponse {
+  dimension: string;
+  srm: number;
+}
+
+interface BayesianDimensionResponse extends BaseDimensionResponse {
+  variations: BayesianVariationResponse[];
+}
+
+interface FrequentistVariationResponse extends BaseDimensionResponse {
+  variations: FrequentistVariationResponse[];
+}
+
+type StatsEngineDimensionResponse =
+  | BayesianDimensionResponse
+  | FrequentistVariationResponse;
+
+export interface ExperimentMetricAnalysis {
+  unknownVariations: string[];
+  multipleExposures: number;
+  dimensions: StatsEngineDimensionResponse[];
+}
